requirements.txt
coverage==4.3.4
dj-database-url==0.4.2
Django==1.10.6
django-cors-middleware==1.3.1
django-crispy-forms==1.6.1
django-filter==1.0.1
django-guardian==1.4.6
django-model-utils==2.6.1
djangorestframework==3.6.2
djangorestframework-jwt==1.8.0
djoser==0.5.4
gunicorn==19.7.0
psycopg2==2.7
PyJWT==1.4.2
pytz==2016.10
requests==2.13.0
requests-mock==1.1.0
six==1.10.0
whitenoise==3.3.0
faker==0.7.17
celery==4.0.2
django-celery-results==1.0.1
